Fixed count ranges Passed

+++ funcs.cpp
@@ -11,13 +11,13 @@ void parseFile(std::istream & input, std::ostream & output){
     std::vector<int> a(n);
     std::copy_n(std::istream_iterator<int>(input),n,a.begin());
 
-    int count=0;
+    unsigned long long int count=0;
     int ir=0;
     for(int il=0; il<a.size()-1; ++il){
         ir = std::max(ir,il+1);
         for(;ir<a.size();++ir){
             if(a[ir]-a[il]>r){
-                count+=a.size()-ir;
+                count+=a.size() - (long)(ir);
                 break;
             }
         }
